feat: white-label rebrand + locale-only URLs (no country segment)
Rebrand the storefront as Kakałowy Sklepik and make Polish the
default locale (no URL prefix — /en gets a prefix, /pl does not).

URL structure: /{country}/{locale}/... -> /{locale}/..., single
fixed default country used server-side for market/currency
resolution instead of a URL segment (single-market store). Product
URLs use the product slug again instead of the ID.

Removed the demo-only Spree/GitHub/quickstart links and logo from
Header, Footer, HeroSection, and the checkout layout — nothing
Spree-branded remains user-visible.
