+++ src/app/[locale]/(checkout)/checkout/[id]/CheckoutSidebar.tsx
