+++ src/app/[locale]/(storefront)/account/forgot-password/page.tsx
