+++ src/app/[locale]/(storefront)/account/register/page.tsx
