+++ src/app/[locale]/(storefront)/account/reset-password/page.tsx
