+++ src/app/[locale]/(storefront)/c/[...permalink]/CategoryBanner.tsx
