+++ src/components/home/HeroSection.tsx
@@ -15,11 +15,6 @@ export async function HeroSection({ basePath, locale }: HeroSectionProps) {
   });
   const storeName = getStoreName();
 
-  /* Demo-only: Remove for production. */
-  const githubUrl = "https://github.com/spree/storefront";
-  const quickstartUrl =
-    "https://spreecommerce.org/docs/developer/getting-started/quickstart";
-
   return (
     <section className="border-b border-gray-200 min-h-[823px] md:min-h-0 flex items-center">
       <div className="container mx-auto px-4 sm:px-6 lg:px-8 py-12 md:py-24">
@@ -33,21 +28,6 @@ export async function HeroSection({ basePath, locale }: HeroSectionProps) {
           <div className="mt-8 flex justify-center gap-4 flex-wrap">
             <Button size="lg" asChild>
               <Link href={`${basePath}/products`}>{t("shopNow")}</Link>
-            </Button>
-            {/* Demo-only: Remove for production. */}
-            <Button variant="outline" size="lg" asChild>
-              <Link href={githubUrl} target="_blank" rel="noopener noreferrer">
-                {t("forkOnGithub")}
-              </Link>
-            </Button>
-            <Button variant="outline" size="lg" asChild>
-              <Link
-                href={quickstartUrl}
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                {t("quickstartGuide")} &rarr;
-              </Link>
             </Button>
           </div>
         </div>
