+++ src/components/layout/Footer.tsx
@@ -8,12 +8,6 @@ import { CurrentYear } from "./CurrentYear";
 const storeName = getStoreName();
 const storeDescription = getStoreDescription();
 
-// Demo-only: Remove for production.
-const githubUrl = "https://github.com/spree/storefront";
-const quickstartUrl =
-  "https://spreecommerce.org/docs/developer/getting-started/quickstart";
-const learnMoreUrl = "https://spreecommerce.org";
-
 interface FooterProps {
   rootCategories: Category[];
   basePath: string;
@@ -32,40 +26,12 @@ export async function Footer({
     <footer className="bg-primary text-gray-300">
       <div className="container mx-auto px-4 sm:px-6 lg:px-8 py-12">
         <div className="grid grid-cols-1 gap-8 md:grid-cols-5">
-          {/* Demo-only: Remove for production. */}
           {/* Brand */}
           <div className="col-span-1 md:col-span-2">
             <span className="text-xl font-bold text-white">{storeName}</span>
             <p className="mt-4 text-sm text-neutral-400">
               {t("description") || storeDescription}
             </p>
-            {/* Demo-only: Remove for production. */}
-            <div className="mt-4 flex flex-col gap-2">
-              <Link
-                href={githubUrl}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="text-sm text-white hover:text-neutral-200 transition-colors font-medium"
-              >
-                {t("forkOnGithub")} &rarr;
-              </Link>
-              <Link
-                href={quickstartUrl}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="text-sm text-neutral-400 hover:text-neutral-200 transition-colors"
-              >
-                {t("quickstartGuide")}
-              </Link>
-              <Link
-                href={learnMoreUrl}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="text-sm text-neutral-400 hover:text-neutral-200 transition-colors"
-              >
-                {t("learnMore")}
-              </Link>
-            </div>
           </div>
 
           {/* Links */}
@@ -150,15 +116,7 @@ export async function Footer({
 
         <div className="mt-8 pt-8 border-t border-neutral-800 text-xs text-neutral-400 text-center">
           <p>
-            &copy; <CurrentYear /> {storeName}. {t("poweredBy")}{" "}
-            <Link
-              href="https://spreecommerce.org"
-              target="_blank"
-              className="text-neutral-400 hover:text-neutral-200 underline transition-colors"
-            >
-              Spree Commerce
-            </Link>{" "}
-            & Next.js.
+            &copy; <CurrentYear /> {storeName}.
           </p>
         </div>
       </div>
