+++ src/lib/metadata/category.ts
@@ -1,40 +1,47 @@
-import type { Metadata } from 'next'
-import { getCachedCategory } from '@/lib/data/cached'
-import { buildCanonicalUrl } from '@/lib/seo'
-import { getStoreUrl } from '@/lib/store'
+import type { Metadata } from "next";
+import { getCachedCategory } from "@/lib/data/cached";
+import { buildCanonicalUrl } from "@/lib/seo";
+import { getStoreUrl } from "@/lib/store";
+import { buildBasePath } from "@/lib/utils/path";
 
 export interface CategoryMetadataParams {
-  country: string
-  locale: string
-  permalink: string[]
+  locale: string;
+  permalink: string[];
 }
 
 export async function generateCategoryMetadata({
-  country,
   locale,
   permalink,
 }: CategoryMetadataParams): Promise<Metadata> {
-  const fullPermalink = permalink.join('/')
+  const fullPermalink = permalink.join("/");
 
-  let category
+  let category;
   try {
-    category = await getCachedCategory(fullPermalink, ['ancestors', 'children'])
+    category = await getCachedCategory(fullPermalink, [
+      "ancestors",
+      "children",
+    ]);
   } catch {
-    return { title: 'Category Not Found' }
+    return { title: "Category Not Found" };
   }
 
   if (!category) {
-    return { title: 'Category Not Found' }
+    return { title: "Category Not Found" };
   }
 
-  const title = category.meta_title || category.name
+  const title = category.meta_title || category.name;
   const description =
-    category.meta_description || category.description || `Browse ${category.name} products.`
+    category.meta_description ||
+    category.description ||
+    `Browse ${category.name} products.`;
 
-  const storeUrl = getStoreUrl()
+  const storeUrl = getStoreUrl();
   const canonicalUrl = storeUrl
-    ? buildCanonicalUrl(storeUrl, `/${country}/${locale}/c/${category.permalink}`)
-    : undefined
+    ? buildCanonicalUrl(
+        storeUrl,
+        `${buildBasePath(locale)}/c/${category.permalink}`,
+      )
+    : undefined;
 
   return {
     title,
@@ -45,14 +52,16 @@ export async function generateCategoryMetadata({
       title,
       description,
       ...(canonicalUrl ? { url: canonicalUrl } : {}),
-      type: 'website',
-      ...(category.image_url ? { images: [{ url: category.image_url, alt: category.name }] } : {}),
+      type: "website",
+      ...(category.image_url
+        ? { images: [{ url: category.image_url, alt: category.name }] }
+        : {}),
     },
     twitter: {
-      card: 'summary_large_image',
+      card: "summary_large_image",
       title,
       description,
       ...(category.image_url ? { images: [category.image_url] } : {}),
     },
-  }
+  };
 }
