+++ src/proxy.ts
@@ -1,9 +1,14 @@
 import { createSpreeMiddleware } from "@/lib/spree/middleware";
-import { getDefaultCountry, getDefaultLocale } from "@/lib/store";
+import {
+  getDefaultCountry,
+  getDefaultLocale,
+  getPrefixedLocales,
+} from "@/lib/store";
 
 export const proxy = createSpreeMiddleware({
   defaultCountry: getDefaultCountry(),
   defaultLocale: getDefaultLocale(),
+  supportedLocales: getPrefixedLocales(),
 });
 
 export const config = {
